Fix filtering of WWW-Authenticate header
When this extension is installed and enabled, visiting a page with HTTP Basic Authentication will result in a request loop.

This comes from the fact that while the extension request
WWW-Authenticate header is defined, it does not check synchronously for
the presence of a PrivateToken auth scheme.

This commit adds a parsing of WWW-Authenticate header to confirm it is a
valid PrivateToken header before redirecting/refreshing the current
webpage.

Closes #6