Add restriction on tab opening to the current active tab
In the existing extension, any tab initiating a PrivateToken request for
a trusted issuer is going to trigger the associated attester tab
opening. This might interrupt the user flow.
This commit introduces an improvement to only allow the current active
tab to open a new tab upon a PrivateToken request.